+++ tasks/release.py
@@ -8,6 +8,7 @@
 from __future__ import print_function
 
 import fileinput
+from builtins import input
 from datetime import date
 
 import yaml
@@ -43,7 +44,7 @@ def push_branches(ctx, force=False):
     version = current_version()
     project_id = cookiecutter_context()['project_id']
     branch_name = 'merge-branch-{}-{}'.format(project_id, version)
-    response = raw_input(
+    response = input(
         'Push local branches to {}? (Y/n) '.format(branch_name)
     )
     if response in ('n', 'N', 'no'):